server/requirements.txt
Adafruit-Blinka==6.9.2
Adafruit-PlatformDetect==3.13.2
Adafruit-PureIO==1.1.8
APScheduler==3.7.0
CacheControl==0.12.6
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
firebase-admin==5.0.0
google-api-core==1.30.0
google-api-python-client==2.9.0
google-auth==1.31.0
google-auth-httplib2==0.1.0
google-cloud-core==1.7.0
google-cloud-firestore==2.1.2
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
grpcio==1.38.0
httplib2==0.19.1
idna==2.10
msgpack==1.0.2
packaging==20.9
pkg-resources==0.0.0
proto-plus==1.18.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyfcm==1.5.1
pyftdi==0.53.1
pyparsing==2.4.7
pyserial==3.5
pytz==2021.1
pyusb==1.1.1
requests==2.25.1
rpi-ws281x==4.2.6
RPi.GPIO==0.7.0
rsa==4.7.2
six==1.16.0
sysv-ipc==1.1.0
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.4
